+++ server/paymentProviders/stripe/index.js
@@ -15,6 +15,7 @@ import models from '../../models';
 
 import creditcard from './creditcard';
 import paymentintent from './payment-intent';
+import { webhook } from './webhook';
 
 const debug = debugLib('stripe');
 
@@ -211,68 +212,5 @@ export default {
     }
   },
 
-  webhook: request => {
-    const requestBody = request.body;
-
-    debug(`Stripe webhook event received : ${request.rawBody}`);
-
-    // Stripe sends test events to production as well
-    // don't do anything if the event is not livemode
-    // NOTE: not using config.env because of ugly tests
-    if (process.env.OC_ENV === 'production' && !requestBody.livemode) {
-      return Promise.resolve();
-    }
-
-    const stripeEvent = {
-      signature: request.headers['stripe-signature'],
-      rawBody: request.rawBody,
-    };
-
-    if (requestBody.type === 'issuing_authorization.request') {
-      return virtualcard.processAuthorization(requestBody.data.object, stripeEvent);
-    }
-
-    if (requestBody.type === 'issuing_authorization.created' && !requestBody.data.object.approved) {
-      return virtualcard.processDeclinedAuthorization(requestBody.data.object, stripeEvent);
-    }
-
-    if (requestBody.type === 'issuing_authorization.updated') {
-      return virtualcard.processUpdatedTransaction(requestBody.data.object, stripeEvent);
-    }
-
-    if (requestBody.type === 'issuing_transaction.created') {
-      return virtualcard.processTransaction(requestBody.data.object, stripeEvent);
-    }
-
-    if (requestBody.type === 'issuing_card.updated') {
-      return virtualcard.processCardUpdate(requestBody.data.object, stripeEvent);
-    }
-
-    /**
-     * We check the event on stripe directly to be sure we don't get a fake event from
-     * someone else
-     */
-    // TODO: Change to https://stripe.com/docs/webhooks/signatures#verify-official-libraries
-    //       to verify the signature without having to make another call to Stripe?
-    return stripe.events.retrieve(requestBody.id, { stripeAccount: requestBody.user_id }).then(event => {
-      if (!event || (event && !event.type)) {
-        throw new errors.BadRequest('Event not found');
-      }
-
-      if (event.type === 'charge.dispute.created') {
-        return creditcard.createDispute(event);
-        // Charge dispute has been closed on Stripe (with status of: won/lost/closed)
-      } else if (event.type === 'charge.dispute.closed') {
-        return creditcard.closeDispute(event);
-      } else if (event.type === 'review.opened') {
-        return creditcard.openReview(event);
-      } else if (event.type === 'review.closed') {
-        return creditcard.closeReview(event);
-      } else {
-        console.log(JSON.stringify(event, null, 4));
-        logger.warn(`Stripe: Webhooks: Received an unsupported event type: ${event.type}`);
-        return;
-      }
-    });
-  },
+  webhook,
 };
